Closing sipXconfig/3.3 branch: svn merge https://scm.sipfoundry.org/rep/sipX/main/sipXconfig@6325 https://scm.sipfoundry.org/rep/sipX/branches/sipXconfig/3.3@6325 .
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@6329 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXconfig/neoconf/src/org/sipfoundry/sipxconfig/search/BeanAdaptor.java
@@ -82,8 +82,7 @@ public interface BeanAdaptor {
 
         public Object transform(Object identity) {
             Identity i = (Identity) identity;
-            // FIXME: remove cast after DataObjectSource is fixed
-            return m_source.load(i.getBeanClass(), (Integer) i.getBeanId());
+            return m_source.load(i.getBeanClass(), i.getBeanId());
         }
     }
 }
